# src/main/java/br/com/rfernandon/demoswaggercodegen/api/PostagemApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package br.com.rfernandon.demoswaggercodegen.api;

import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import br.com.rfernandon.demoswaggercodegen.model.ErrorDetails;
import br.com.rfernandon.demoswaggercodegen.model.Postagens;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-03-03T23:00:40.757-03:00")

@Api(value = "Postagem", description = "Lista de endpoints disponíveis para o recurso", tags={ "Postagem" })
@RequestMapping(value = "/api/v1")
public interface PostagemApi {

    @ApiOperation(value = "Cadastrar publição", nickname = "postagemPost", notes = "Este endpoint cadastra uma nova publição", response = Postagens.class, authorizations = {
        @Authorization(value = "api_basic_auth")
    }, tags={ "Postagem", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "OK - Retorna a publição cadastrada", response = Postagens.class),
        @ApiResponse(code = 400, message = "Bad Request - Requisição inválida", response = ErrorDetails.class),
        @ApiResponse(code = 500, message = "Internal Server Error - Ocorreu um erro inesperado no servidor", response = ErrorDetails.class) })
    @RequestMapping(value = "/postagem",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    Mono<Postagens> postagemPost(@ApiParam(value = "Parâmetro referente a publição" ,required=true )  @Valid @RequestBody Postagens publicacao);

}